requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
comtypes==1.4.5
cryptography==43.0.0
cssselect2==0.7.0
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.2
django-allauth==64.0.0
django-ckeditor==6.7.0
django-ckeditor-5==0.2.13
django-default-language==0.0.1
django-filer==3.1.4
django-grappelli==3.0.9
django-js-asset==2.2.0
django-mptt==0.16.0
django-parler==2.3
django-polymorphic==3.1.0
django-rosetta==0.10.0
EbookLib==0.18
fire==0.6.0
fonttools==4.53.1
fpdf==1.7.2
idna==3.7
img2pdf==0.5.1
lxml==5.2.2
mysqlclient==2.2.4
numpy==2.0.0
oauthlib==3.2.2
opencv-python-headless==4.10.0.84
packaging==24.0
pdf2docx==0.5.8
pdf2image==1.17.0
pikepdf==8.15.1
pillow==10.3.0
polib==1.2.0
pptxtopdf==0.0.2
pycparser==2.22
PyJWT==2.9.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyPDF2==3.0.1
pytesseract==0.3.10
python-docx==1.1.2
python-pptx==0.6.23
python3-openid==3.2.0
reportlab==4.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.5.0
svglib==1.5.1
termcolor==2.4.0
tinycss2==1.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.7.0
wrapt==1.16.0
XlsxWriter==3.2.0
